+++ photutils/detection/irafstarfinder.py
@@ -11,7 +11,7 @@ from astropy.utils.decorators import deprecated_renamed_argument
 from astropy.utils.exceptions import AstropyDeprecationWarning
 
 from photutils.detection.core import (StarFinderBase, StarFinderCatalogBase,
-                                      _StarFinderKernel, _validate_brightest)
+                                      _StarFinderKernel, _validate_n_brightest)
 from photutils.utils._convolution import _filter_data
 from photutils.utils._parameters import warn_positional_kwargs
 from photutils.utils._quantity_helpers import isscalar, process_quantities
@@ -72,9 +72,9 @@ class IRAFStarFinder(StarFinderBase):
         the convolution kernel from the image borders. The default is
         `False`, which is the mode used by starfind.
 
-    brightest : int, None, optional
+    n_brightest : int, None, optional
         The number of brightest objects to keep after sorting the source
-        list by flux. If ``brightest`` is set to `None`, all objects
+        list by flux. If ``n_brightest`` is set to `None`, all objects
         will be selected.
 
     peak_max : float, None, optional
@@ -158,10 +158,11 @@ class IRAFStarFinder(StarFinderBase):
     """
 
     @warn_positional_kwargs(since='3.0', until='4.0')
+    @deprecated_renamed_argument('brightest', 'n_brightest', '3.0')
     @deprecated_renamed_argument('peakmax', 'peak_max', '3.0')
     def __init__(self, threshold, fwhm, sigma_radius=1.5, minsep_fwhm=2.5,
                  sharpness_range=(0.5, 2.0), roundness_range=(0.0, 0.2),
-                 exclude_border=False, brightest=None, peak_max=None,
+                 exclude_border=False, n_brightest=None, peak_max=None,
                  xycoords=None, min_separation=None,
                  sharplo=_NODEFAULT, sharphi=_NODEFAULT,
                  roundlo=_NODEFAULT, roundhi=_NODEFAULT):
@@ -220,7 +221,7 @@ class IRAFStarFinder(StarFinderBase):
         self.sharpness_range = sharpness_range
         self.roundness_range = roundness_range
         self.exclude_border = exclude_border
-        self.brightest = _validate_brightest(brightest)
+        self.n_brightest = _validate_n_brightest(n_brightest)
         self.peak_max = peak_max
 
         if xycoords is not None:
@@ -245,7 +246,7 @@ class IRAFStarFinder(StarFinderBase):
     def _repr_str_params(self):
         params = ('threshold', 'fwhm', 'sigma_radius', 'minsep_fwhm',
                   'sharpness_range', 'roundness_range',
-                  'exclude_border', 'brightest', 'peak_max', 'xycoords',
+                  'exclude_border', 'n_brightest', 'peak_max', 'xycoords',
                   'min_separation')
         overrides = {}
         if not isscalar(self.threshold):
@@ -309,7 +310,7 @@ class IRAFStarFinder(StarFinderBase):
                                       self.kernel,
                                       sharpness_range=self.sharpness_range,
                                       roundness_range=self.roundness_range,
-                                      brightest=self.brightest,
+                                      n_brightest=self.n_brightest,
                                       peak_max=self.peak_max)
 
     @warn_positional_kwargs(since='3.0', until='4.0')
@@ -400,9 +401,9 @@ class _IRAFStarFinderCatalog(StarFinderCatalogBase):
         detection. Objects with roundness outside this range will be
         rejected.
 
-    brightest : int, None, optional
+    n_brightest : int, None, optional
         The number of brightest objects to keep after sorting the source
-        list by flux. If ``brightest`` is set to `None`, all objects
+        list by flux. If ``n_brightest`` is set to `None`, all objects
         will be selected.
 
     peak_max : float, None, optional
@@ -417,7 +418,7 @@ class _IRAFStarFinderCatalog(StarFinderCatalogBase):
 
     def __init__(self, data, convolved_data, xypos, kernel, *,
                  sharpness_range=(0.2, 1.0), roundness_range=(-1.0, 1.0),
-                 brightest=None, peak_max=None):
+                 n_brightest=None, peak_max=None):
 
         # Validate the units, but do not strip them
         inputs = (data, convolved_data, peak_max)
@@ -425,7 +426,7 @@ class _IRAFStarFinderCatalog(StarFinderCatalogBase):
         _ = process_quantities(inputs, names)
 
         super().__init__(data, xypos, kernel,
-                         brightest=brightest,
+                         n_brightest=n_brightest,
                          peak_max=peak_max)
 
         self.convolved_data = convolved_data
@@ -441,7 +442,7 @@ class _IRAFStarFinderCatalog(StarFinderCatalogBase):
         Return a tuple of attribute names to copy during slicing.
         """
         return ('data', 'unit', 'convolved_data', 'kernel',
-                'sharpness_range', 'roundness_range', 'brightest',
+                'sharpness_range', 'roundness_range', 'n_brightest',
                 'peak_max', 'cutout_shape', 'default_columns')
 
     @lazyproperty